clone_or_pull_repos/requirements.txt
tqdm>=4.0.0
python-decouple